DesignCSS, Rename: Home.jsx와 Search.jsx 디자인 수정 및 파일명 변경
