+++ src/pages/Search.jsx
@@ -1,7 +1,18 @@
 import React, { useEffect, useState } from 'react';
-import TopSearchNav from './nav/TopSearchNav';
+import TopSearchNav from '../components/nav/TopSearchNav';
 import axios from 'axios';
-import UserSearch from './UserSearch';
+import UserSearch from '../components/UserSearch';
+import Nav from '../components/nav/Nav';
+import TabMenu from '../components/tab/TabMenu';
+import styled from 'styled-components';
+
+const Ul = styled.ul`
+    margin: 60px 16px;
+`;
+
+const Li = styled.li`
+    padding-top: 16px;
+`;
 
 function Search() {
     const [keyword, setKeyword] = useState('');
@@ -19,17 +30,6 @@ function Search() {
                 'Authorization': `Bearer ${token}`,
                 'Content-type': 'application/json',
             },
-            data: [
-                {
-                    _id: String,
-                    username: String,
-                    accountname: String,
-                    following: [],
-                    follower: [],
-                    followerCount: Number,
-                    followingCount: Number,
-                },
-            ],
         })
             .then((response) => setUsers(response))
             .then(console.log(users));
@@ -41,21 +41,22 @@ function Search() {
     };
     return (
         <>
-            <TopSearchNav propFunc={inputOnChange} />
-            <ul>
+            <Nav>
+                <TopSearchNav propFunc={inputOnChange} />
+            </Nav>
+            <Ul>
                 {users.data &&
                     users.data.map((user) => (
-                        <li key={user._id}>
-                            {/* {user.username} ({user.accountname}) */}
+                        <Li key={user._id}>
                             <UserSearch
                                 userImg={user.image}
                                 username={user.username}
                                 accountname={user.accountname}
                             />
-                        </li>
+                        </Li>
                     ))}
-            </ul>
-            <ul></ul>
+            </Ul>
+            <TabMenu img={'homeImg'} />
         </>
     );
 }
